requirements.txt
aiofiles==23.2.1
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
blinker==1.6.2
build==1.1.1
CacheControl==0.14.0
cachetools==5.3.1
certifi==2022.12.7
cffi==1.16.0
chardet==5.1.0
charset-normalizer==3.1.0
chromadb==0.3.29
cleo==2.1.0
click==8.1.5
clickhouse-connect==0.6.6
colab-env==0.2.0
coloredlogs==15.0.1
contourpy==1.1.1
crashtest==0.4.1
cryptography==41.0.2
cycler==0.12.0
dashscope==1.19.3
dataclasses-json==0.5.8
decorator==5.1.1
distlib==0.3.8
docarray==0.34.0
duckdb==0.8.1
dulwich==0.21.7
et-xmlfile==1.1.0
fastapi==0.85.1
fastjsonschema==2.19.1
ffmpy==0.3.1
filelock==3.12.4
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.43.0
frozenlist==1.3.3
fsspec==2023.9.2
gitdb==4.0.10
GitPython==3.1.32
gradio==3.47.1
gradio_client==0.6.0
greenlet==2.0.2
h11==0.14.0
hnswlib==0.7.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
install==1.3.5
installer==0.7.0
jaraco.classes==3.3.1
Jinja2==3.1.2
joblib==1.3.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
keyring==24.3.1
kiwisolver==1.4.5
langchain==0.0.232
langchain-community==0.0.32
langchain-core==0.1.41
langchainplus-sdk==0.0.17
langsmith==0.1.42
lxml==4.9.3
lz4==4.3.2
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.8.0
mdurl==0.1.2
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
msg-parser==1.2.0
msgpack==1.0.8
multidict==6.0.4
mypy-extensions==1.0.0
neo4j==5.19.0
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.0
olefile==0.46
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
orjson==3.10.0
overrides==7.3.1
packaging==23.2
pandas==2.0.3
pdf2image==1.16.3
pdfminer.six==20221105
pexpect==4.9.0
Pillow==9.5.0
pkginfo==1.10.0
platformdirs==4.2.0
podman-compose==1.0.3
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.6.0
posthog==3.0.1
protobuf==4.23.4
ptyprocess==0.7.0
pulsar-client==3.2.0
pyarrow==12.0.1
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.9
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.15.1
Pympler==1.0.1
pypandoc==1.11
pyparsing==3.1.1
pypdf==3.11.1
PyPDF2==3.0.1
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.0
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
rapidfuzz==3.6.2
referencing==0.29.1
regex==2023.6.3
requests==2.29.0
requests-toolbelt==1.0.0
rich==13.4.2
rpds-py==0.8.10
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.17
starlette==0.20.4
streamlit==1.25.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
tomlkit==0.12.4
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
trove-classifiers==2024.3.3
types-requests==2.31.0.1
types-urllib3==1.26.25.13
typing-inspect==0.9.0
typing_extensions==4.7.0
tzdata==2023.3
tzlocal==4.3.1
unstructured==0.8.1
urllib3==1.26.15
uvicorn==0.22.0
uvloop==0.17.0
validators==0.20.0
virtualenv==20.25.1
watchdog==3.0.0
watchfiles==0.19.0
websockets==11.0.3
xattr==1.1.0
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.16.1
zstandard==0.21.0
